+++ src/App.js
@@ -1,4 +1,12 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { Routes, Route } from 'react-router-dom';
+
+import { setCurrentUser } from './store/user/user.action';
+import {
+  onAuthStateChangedListener,
+  createUserDocFromAuth
+} from './utils/firebase/firebase.utils';
 
 import { Home } from './routes/home/home.component';
 import { Navigation } from './routes/navigation/navigation.component';
@@ -7,6 +15,20 @@ import { Shop } from './routes/shop/shop.component';
 import { Checkout } from './routes/ceckout/checkout.component';
 
 const App = () => {
+  const dispatch = useDispatch();
+  useEffect(() => {
+    const unsubscribe = onAuthStateChangedListener(user => {
+      if (user) {
+        createUserDocFromAuth(user);
+      }
+      dispatch(setCurrentUser(user));
+    });
+
+    return unsubscribe;
+  }, []);
+  // ESLint gives mistake, because it see dispatch function inside.
+  // 'dispatch' may be put inside to hide error. But it may cause missunderstanding.
+  // It seems it will rerun every time dispatch fires, but redux logic is that it woul fire once.
   return (
     <Routes>
       <Route path='/' element={<Navigation />}>
