+++ src/routes/authentication/authentication.component.jsx
@@ -1,13 +1,13 @@
-import { useContext } from 'react';
 import { Navigate } from 'react-router-dom';
-import { UserContext } from '../../contexts/user.context';
+import { useSelector } from 'react-redux';
 
+import { selectCurrentUser } from '../../store/user/user.selector';
 import { SignInForm } from '../../components/sign-in/sign-in.component';
 import { SignUpForm } from '../../components/sign-up-form/sign-up-form.component';
 import { Container } from './authentication.styles';
 
 export const Authentication = () => {
-  const { currentUser } = useContext(UserContext);
+  const currentUser = useSelector(selectCurrentUser);
 
   return (
     <Container>
