+++ src/routes/navigation/navigation.component.jsx
@@ -1,18 +1,19 @@
 import { useContext } from 'react';
 import { Outlet } from 'react-router-dom';
+import { useSelector } from 'react-redux';
 
 import { CartIcon } from '../../components/cart-icon/cart-icon.component';
 import { CartDropdown } from '../../components/cart-dropdown/cart-dropdown.component';
 
+import { selectCurrentUser } from '../../store/user/user.selector';
 import { ReactComponent as CrownLogo } from '../../assets/crown.svg';
-import { UserContext } from '../../contexts/user.context';
 import { CartContext } from '../../contexts/cart.context';
 import { signOutUser } from '../../utils/firebase/firebase.utils';
 
 import { Nav, LogoLink, Links, NavLink } from './navigation.styles';
 
 export const Navigation = () => {
-  const { currentUser } = useContext(UserContext);
+  const currentUser = useSelector(selectCurrentUser);
   const { isCartOpen } = useContext(CartContext);
 
   return (
