+++ src/store/root.reducer.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+
+import { userReducer } from './user/user.reducer';
+
+export const rootReducer = combineReducers({
+  user: userReducer
+});
