+++ src/store/store.js
@@ -1,0 +1,9 @@
+import { compose, createStore, applyMiddleware } from 'redux';
+import logger from 'redux-logger';
+
+import { rootReducer } from './root.reducer';
+
+const middleWares = [logger];
+const composedEnhancers = compose(applyMiddleware(...middleWares));
+
+export const store = createStore(rootReducer, undefined, composedEnhancers);
